New util function: promise queue
