+++ src/utils/index.ts
@@ -1,8 +1,8 @@
 import * as CheckAddress from './check-address';
 import * as ToTimestamp from './timestamp';
 import ThrowError from './throw-error';
-import { Request, Queue, WrapperRequest } from './request';
+import { Request, PromiseQueue, WrapperRequest } from './request';
 import { Configuration } from './configuration';
 import * as Checkers from './utilities';
 
-export { CheckAddress, ToTimestamp, ThrowError, Request, Configuration, Checkers, Queue, WrapperRequest };
+export { CheckAddress, ToTimestamp, ThrowError, Request, Configuration, Checkers, PromiseQueue, WrapperRequest };
